:sparkles: Accummulate usecase implemented.

+++ internal/domain/transactions/model/transaction.go
@@ -12,7 +12,7 @@ type Transaction struct {
 	ID           *uuid.UUID `json:"id" gorm:"type:uuid;default:uuid_generate_v4();primaryKey"`
 	BranchID     *uuid.UUID `json:"branch_id" gorm:"type:uuid;not null"`
 	UserID       *uuid.UUID `json:"user_id" gorm:"not null"`
-	Amount       int        `json:"amount" gorm:"not null"`
+	Amount       float64    `json:"amount" gorm:"not null"`
 	PurchaseDate time.Time  `json:"purchase_date" gorm:"not null"`
 	Accumulated  bool       `json:"accumulated" gorm:"not null;default:false;index"`
 	Redeemed     bool       `json:"redeemed" gorm:"not null;default:false;index"`
